Fix hover state of key navigation buttons
We were passing the first top-level channel to the widget but that was incorrect because it's not hoverable, and even if it was, it wouldn't be the thing that is hovered when the mouse cursor is on the widget. The widget is on the track item, so that's what should be passed.

Also, when clicking on the "add key" button and next/previous key buttons, we're not supposed to only look at the first top-level channel. If there are multiple top-level channels, we should key all of them (as applicable) if we click the "add key" button, and we should take into account all of their keys to navigate to the next/previous ones.

#jira none
#rb andrew.rodham
#preflight 62ab7bf0ae3f2bad9adcb0f2

[CL 20692273 by ludovic chabant in ue5-main branch]