Fix a silent inlined depenceny in the RemoteSession and Bridge plugins.
They both implement their own MessageHandler, but don't depend on ApplicationCore. The only reason that this would compile before was because of the base GenericApplicationMessageHandler being inlined into their files. If we change anything in the GenericMessageHandler then we get linker errors. We will need to reference the InputDeviceMapper in there, so this is a necessary change.

#jira none
#rb trivial
#rnx
#preflight 62b60c8bfb070eb2f45ce527

[CL 20815256 by ben hoffman in ue5-main branch]