ImgMedia and FilePathPicker: Added option for the file picker to return an absolute path when the path comes from the system dialog.
This eliminates the ambiguity when the path is returned is relative, since for the image sequence path we support paths to be specified relative to the project (typed by the user) while the dialog returned paths relative to the process, but the callback for the textfield and the dialog are the same at the file picker layer. The issue ultimately happens because when ".." hits the root folder, it behaves as a no-op. So multiple relative paths can point to the same file for the same given starting directory.

The flag added to SFilePathPicker defaults to the previous behavior (which is to return paths relative to the process).

#rb trivial
#jira UE-157535
#preflight 62b611d40716a4b4d12854b0

[CL 20815496 by Alejandro Arango in ue5-main branch]