This change adds a UBodySetup from-text factory called FBodySetupObjectTextFactory, which is used in PHAT and the Static Mesh editor to enable cross-copy/pasting of shapes from physics setups.
In FN, the workflow for creating physics assets for vehicles often involves importing at least one static mesh and copying collision shapes from it. It's currently really challenging to copy individual shapes between static and skeletal mesh body setups. With the added shape copy/paste commands, we get a lot more flexibility out of SM editor and PHAT, and a somewhat simpler workflow.

If we have physics assets A and B and we want to add one shape (or a specific subset of shapes) from a body in A to a body in B, this change is needed to do that. Since I'm trying to do this all the time while editing physics assets for vehicles, I use this shelf locally.

#rb benn.gallagher, nick.brett, satchit.subramanian

#ROBOMERGE-AUTHOR: steven.barnett
#ROBOMERGE-SOURCE: CL 20861874 via CL 20861887 via CL 20861898
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v971-20777995)

[CL 20876607 by steven barnett in ue5-main branch]