Sequencer: Added the ability to set per-frame budgets for sequencer evaluations.
Setting a budget will cause evaluation to become time-sliced across multiple frames, as soon as the supplied budget is exceeded.
By default this is turned off, but can be overridden by specifying an Evaluation Budget on the playback settings.
Mismatches of budgets will cause the smallest budget to be used, but will emit a warning if budgeted and non-budgeted evaluations are defined for the same tick interval.

Implementation details:

The evaluation loop is set up as a crude state machine which allows us to break out of the
loop and pick up where we left off next frame. Each bit within the FlushState bitmask defines
a particular task that needs to execute. In this way we can loop the evaluation while there
are updates pending by simply setting the necessary flags for each task. For convenience, the
required tasks to perform another iteration are defined in ERunnerFlushState::LoopEval.

#rb Ludovic.Chabant, Lonnie.Li
#preflight 62bc7d7c115d146ce3d339f0

#ROBOMERGE-AUTHOR: andrew.rodham
#ROBOMERGE-SOURCE: CL 20877297 via CL 20877416 via CL 20877526
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v971-20777995)

[CL 20880134 by andrew rodham in ue5-main branch]