Switchboard: add a setting to allow finding levels that are inside project plugins
This adds a "maps_plugin_filters" StringListSetting that can be used to add
levels located inside plugins to the "Levels" dropdown menu in the main
Switchboard dialog. The setting is a list of filters similar to the existing "maps_filter"
setting which tests against map paths, but "maps_plugin_filters" is tested against
plugin names. Any plugins whose name matches any of the filters will be searched
for levels when populating the dropdown menu.

#jira UE-158000
#rb jason.walter
#preflight skip

[CL 20880258 by matt johnson in ue5-main branch]