Fixed deadlock if FLongPackagePathsSingleton isn't yet initialised.
The constructor for FLongPackagePathsSingleton would take a write lock. This would block however as there are Get() call sites where a read lock is already being held. But as the singleton construction is thread-safe there is no need to take a write lock in constructor and it has been removed. To mitigate the error returning in the future the Get() is called before taking any read locks. The lock was renamed and moved into FLongPackagePathsSingleton instead of a global.

#rb cm
#preflight 62be97728d5e67875907ebcf

[CL 20911412 by Martin Ridgers in ue5-main branch]