UE-151338 - Implement refactored implicit conversion for math types.
Added support for FVector2D and FBox2D. Unfortunately, this is a bit of a hack. We currently don't support UScriptStruct creation for FVector2d and FBox2d, since their FNames collide with FVector2D and FBox2D, respectively.

We can work around this by simply not relying on a UScriptStruct double variant to exist. The logic for the conversion table can simply fall back to use the base UScriptStruct instead. It's a little ugly, but it at least allows implicit conversion between these types.

Additionally, the "varianttype1" and "varianttype2" names seemed a bit too abstract. These were changed into "doublevariant" and "floatvariant", which helps with readability.

#jira UE-151338
#rb andrew.davidson
#preflight 62c460caf514983340a69c1a

[CL 20944842 by dave jones2 in ue5-main branch]