Lumen - replace cone skylight evaluation with a ray sample.
There's no point in having cone evaluation only for a single place in the pipeline. Additionally this was causing a lot of leaking with bright skylights, especially in radiosity which trace very wide cones. E.g. when a ray hits black lower sky hemisphere it still could bring in color from the upper one due to the ray footprint.

[FYI] Daniel.Wright

#ROBOMERGE-AUTHOR: krzysztof.narkowicz
#ROBOMERGE-SOURCE: CL 20969385 via CL 20973166 via CL 20978220
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v972-20964824)

[CL 20979993 by krzysztof narkowicz in ue5-main branch]