The error tests in "test_type_hints" are now being absorbed by the test framework.
"test_type_hints" contains several tests that check if error messages can be created through Python. These tests however fail on Horde because due to an error being reported. To fix this the tests are enabled to absorb these errors and check if they're called exactly once on the test. When running these calls outside the test framework, the additional calls are no-ops for all intends and purposes.

#jira: UE-155636
#preflight 62c847b9f671e8d2fad0bde8

[CL 21008344 by ronald koppers in ue5-main branch]