Fix crash after rebuilding lighting if some materials don't have the "used with static lighting" flag set.
FShaderCompilingManager::ProcessCompiledShaderMaps was storing smart pointers to material resources in a local map, which meant that they had a refcount of 2 when UMaterialInstance::CheckMaterialUsage found the missing flag and tried to recreate the resources. This in turn made the FMaterial destructor trigger an assertion failure, because the refcount was still 1 after removing the resource from the shader map. Emptying the local map before recreating the render state fixes the issue.

#jira UE-158279
#rnx
#preflight https://horde.devtools.epicgames.com/job/62c714882f2d04691854d552
#rb Jason.Nadro

[CL 21037970 by mihnea balta in ue5-main branch]