OpenXR + VP: Restore Editor Viewport correctly after exiting VR Editor Mode (SIMPLE)
OpenXR relies on the logic in FSceneViewport::OnDrawViewport to handle Viewport resizes when stereo is active. Previously, I had a change that only went thru
for PIE + Stereo (VRPIE). I need the same logic for VREM, so I can change the PIE check to just Editor. The viewport handles the check for active stereo rendering already.

#jira UE-150036
#rb Jeff.Fisher Zach.Brockway
#preflight 62cdb5a38e451a6513727af3

[CL 21059582 by robert srinivasiah in ue5-main branch]