Sequence Recorder: Hide Sequence Recorder from menu. Can be re-enabled with cvar "SequenceRecorder"
For 5.1, this just hides the Sequence Recorder from the menu so you can't bring it up. There is already a big red bar in the UI that says it will be removed at some point. For 5.2, we will remove the Sequence Recorder UI entirely. This doesn't change anything about recording from Persona.

#rb ludovic.chabant
#preflight 62d049178a3c9910d1499ff1
#jira none

[CL 21093281 by Max Chen in ue5-main branch]