Static Mesh DI respects per component painted vertex colors.
TODO: Some callback from the Vertex Painting / Static Mesh to update Niagara when there are changes to the vertex overrides or they're removed etc.
TODO: Verify lifetime of OverrideVertexColors passed to GPU sims. Unclear atm if this is safe. It is safe for an RT command to use this as the destruction is done RT safe. However, holding onto the ref beyond a single RT command as Niagara does seems unsafe. Though the skeleltal mesh DI also does this so I may be missing some safety mechanism here.

#rb Rob.Krajcarski
[FYI] Wyeth.Johnson

#ROBOMERGE-OWNER: jaren.peterson
#ROBOMERGE-AUTHOR: simon.tovey
#ROBOMERGE-SOURCE: CL 21268026 via CL 21268083 via CL 21268091
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v972-20964824)

[CL 21270136 by simon tovey in ue5-main branch]