Multi-view Lumen lighting artifacts:  CVar added to disable feature that only updates ray tracing scene once.  Costs around 1.6 ms GPU for a Virtual Production test case (no change in perf outside nDisplay).  The optimization is new to 5.1, so this is just reverting to 5.0 perf levels.  Planning to investigate more, but wanted artifacts solved in the meantime for Virtual Production clients.  The CVar would be useful regardless for A/B testing.
It's possible the artifacts are caused by a bug, or the ray tracing scene needs to be upgraded in some way to be mul0ti-view aware.  If neither of those options work, we could consider supporting multiple ray tracing scenes.  If views are already on separate GPUs, that would involve no additional cost.

#jira UE-159957
#rnx
#rb jason.nadro
#preflight 62f288c3461e862154ff0fc1

[CL 21293337 by jason hoerner in ue5-main branch]