fix selection outline for HDR editor: ScRGB can generate negative values
CL 20982761 didn't work because sign(x) = 0 when x=0, meaning an outline on a black pixel would not work exactly the same. Instead, we now leave the color as-is when no outline is detected, and remove negative values when applying the outline. The result is technically incorrect (ie we should linearize properly depending on the gamma encoding of the backbuffer and have everything work in the same color space), but this allows for a simple fix

#preflight 62d090653b057e72746b252b
#jira UE-151653
#rb eric.renaudhoude

[CL 21422112 by benjamin rouveyrol in ue5-main branch]