Sequencer: Allow reusing an animation asset when baking to it.
Added an argument to ExecuteNewAnimAsset, bAllowReplaceExisting, which when true will look for an existing asset and prompt the user whether to use that asset as opposed to calling create asset. When baking to an animation asset, this allows the user to bake to an existing animation asset. Previously, it would try to create the same named asset but would fail because the asset may be actively referenced.

#jira UE-160925
#preflight 62fe63843f75602b2f0655ea
#rb thomas.sarkanen, mike.zyracki

[CL 21459595 by Max Chen in ue5-main branch]