Fixed issue regarding the refraction attribute when Make/Set MaterialAttributes nodes are used.
Issue was caused by error in logic that detects whether the material applies distortion. Old logic was checking only for terminal output node to write to the Refraction attribute, while the node that write to the Refraction attribute might be somewhere in the subgraph that eventaully leads to that terminal node. The solution was to look traverse the entire subgraph looking for any node that sets the Refraction attribute.

#jira UE-161541
#preflight 6304910d5366f61a42605ddc

[CL 21503492 by massimo tristano in ue5-main branch]