Fix the MultipleLockTexture2D RHI unit test
This test was trying to read back the contents of a texture using the Lock function, which is only supported on a few RHIs. The correct way to do it is to copy to a CPU-accessible texture and use MapStagingSurface on that. Added a check in the Vulkan implementation of RHILockTexture2DArray to catch invalid usage (RHILockTexture2D already had this check).

Also fixed a typo in a couple of RHI unit test messages, and made the MultipleLockTexture2D report success/failure to the log.

#rnx
#jira none
#rb trivial
#fyi Richard.Wallis
#preflight skip

[CL 21677035 by mihnea balta in ue5-main branch]