Automation: Fixed staging of blueprint-only projects on some platforms.
The problem was that blueprint-only projects would return true for IsCodeBasedUProjectFile during staging. The issue was triggered by CL 21214930 where PropertyCacheKey (ProjectUtils.cs) was added. Before this change GetProjectProperties worked based only on the path to .uproject. After the change, it also takes the platform and configurations into account. It means that:

GetProjectProperties("...", SomePlatform, Debug)
GetProjectProperties("...", null, Debug)

may return different results. While it's probably the core or the fix this change implemented, it may given unexpected results because when null is passed as a list of platforms, internally all available platforms are taken into account and if only one of them requires the project to be compiled IsCodeBasedUProjectFile will return true, which may not be what our code is expected to see. That's what happened during staging - IsCodeBasedUProjectFile was invoked two times in a slightly different way leading to the automation code handling blueprint-only projects in an inconsistent-way during the whole process.

Fixed by modifying DeploymentContext to initialize IsCodeBasedProject by passing both the current platform and configurations.

#preflight 630e6bd0501b64ba334c3c2a
#rb none
#jira UE-162001
#rnx

[CL 21711241 by Wojciech Krywult in ue5-main branch]