Expose the number of frames of delay before taking a screenshot
The code had a hardcoded 5 frame delay. All this patch does is expose a setting to control this value (with the default being 5 for backwards compatibility).

The logic is unchanged, first the time based delay is checked, then the frame based delay is checked. You can set either delay to 0 to only use one of them.

This new feature will be used in path tracing related tests to ensure proper convergence regardless of the speed of the machine running the test and without having to set overly lengthy time delays which slow down faster machines without providing guarantees on slower machines.

I have verified that the feature works as expected on all the path tracing tests, including the ones that were frequently failing due to being on the threshold of running too slowly to accumulate enough samples by the time delay.

#jira UE-159434
#rb Jerome.Delattre
#preflight 630fb989556fc14dce9f35a5

[CL 21726693 by chris kulla in ue5-main branch]