Sequencer: Check for valid player
The fixed some crashes that occurred when saving a level sequence while in a multi-user session. I don't have a simple repro to test against but this was tested in the Partner-Otter branch. It's possible that the PlayerUseCounts are being reset or not even being added when it gets to here.

#jira UE-159701
#rb ludovic.chabant
#preflight 62d72d6e164251d065f6eecf

#ushell-cherrypick of 21168439 by Max.Chen

[CL 21771314 by ludovic chabant in ue5-main branch]