Add the bShouldOnlyTriggerLastActionInChord option to Enhanced Input, defaulted to on. This will track input actions that have Chorded Triggers on them, and allow the filtering of triggered delegates that are fired based on which chorded actions are being triggered this tick.
This will allow you to have a chorded action trigger separately while ebing bound to the same key as another, non-chorded action. For example, you can have Left Trigger fire on pressed, right trigger fire on pressed, and have a Chorded action for when they are triggered together.

The example above used to result in the Left trigger being fired at the same time as the chorded trigger, making it difficult to make intuitive chords. Now, only the chorded action will be fired.

#jira UE-149152
#preflight 6317578a032bd43d86489e7f
#rb justin.hare
#rb david.garza

[CL 21817709 by ben hoffman in ue5-main branch]