Scene Capture:  Remove arbitrary Scene Capture 10x memory estimate provided in large render target warning in Texture Editor.  We now have a computed Scene Capture memory estimate displayed in the Texture Editor which is far more accurate, and the number provided in the warning is meaningless and confusing (got a UDN asking about the number).
The actual number in my tests was only 1.17x, but the number can go up as you add rendering features, and the overhead multiplier also varies with the bit depth of the render target.  For example, turning on TAA, which defaults to off, does create a 10x overhead, and Lumen on top of that can bump that to 20x or more.  And, if you had an 8-bit output, the output render target gets smaller, but the persistent data stays the same size (the scene renderer itself doesn't switch to monochrome for any of its persistent data -- it just blits to the lower bit depth at the end), so the overhead relative to the target's memory goes up (1.68x for the simple case with no expensive features).

#jira none
#rnx
#rb jason.nadro
#preflight 63179417ec45fbf3d7d0d8ce

[CL 21855738 by jason hoerner in ue5-main branch]