Explicitly allocate pages for hook trampolines.
Hooking a Windows API works by patching in a "jmp" instruction. The instructions that are overwritten by this patch are moved elsewhere (and appended with another jmp) such that the original function can still be called - I called this a trampoline. Prior to this change, trampolines were written into the null-filled slack space at the end of the code section. This however turned out to be too small as more hooks were added and as Windows' DLLs mutated. This change instead uses a page of address space which it allocates just before the DLL being patched (proximity is important as the jmp instruction used only has 32 bit signed displacement).

#rb im
#rnx
#jira UE-160315 UE-163060
#rb cdp
#preflight 63188f4fd135b61bc5ad67d5

#ushell-cherrypick of 21421841 by Martin.Ridgers

[CL 21864593 by catalin dragoiu in ue5-main branch]